Here new non working version
Still in debug

+++ processing/sigmadeltamodulator.hpp
@@ -4,7 +4,7 @@
 #include "const.hpp"
 
 
-typedef uint8_t SDMOutputType;
+typedef uint32_t SDMOutputType;
 
 class SigmaDeltaModulator{
